+++ app/lib/authentication/types/common.ts
@@ -1,11 +1,11 @@
-import type { User } from "./user.server";
+import type { UserSchema } from "./user";
 
 export interface AuthContextValue {
-  user: User | null;
+  user: UserSchema | null;
   isAuthenticated: boolean;
 }
 
 export interface AuthProviderProps {
   children: React.ReactNode;
-  user: User | null;
+  user: UserSchema | null;
 }